backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.6
requests>=2.31.0
numpy>=1.26.0
python-dotenv>=1.0.0
python-docx>=1.1.0
pymupdf>=1.24.0
pdfminer.six>=20221105
sentence-transformers>=2.2.2
pydantic>=2.10.0
pytest>=7.4.0
httpx>=0.25.0
google-genai>=1.0.0
